[REFACTOR] [Hacker Rank] Warm Up problems moved to their own directory / package.
