+++ algorithm-exercises-java/src/main/java/ae/hackerrank/warmup/CompareTriplets.java
@@ -1,4 +1,4 @@
-package ae.hackerrank;
+package ae.hackerrank.warmup;
 
 import java.util.Arrays;
 import java.util.List;
