+++ algorithm-exercises-java/src/main/java/ae/hackerrank/warmup/PlusMinus.java
@@ -1,4 +1,4 @@
-package ae.hackerrank;
+package ae.hackerrank.warmup;
 
 import java.util.ArrayList;
 import java.util.List;
