+++ algorithm-exercises-java/src/main/java/ae/hackerrank/warmup/SolveMeFirst.java
@@ -1,4 +1,4 @@
-package ae.hackerrank;
+package ae.hackerrank.warmup;
 
 /**
  * Solve Me First.
